+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- `scalr_run_trigger`: add missing import support
+
 ## [3.16.2] - 2026-05-05
 
 ### Fixed